Fix "exceeded available parameter key space" bug with large POST requests
Rack attempts to parse the POST data as a form if the Content-Type header
matches one of its understood formats. Sometimes the Content-Type header
is not set correctly by calls to the API and it tries to parse large requests.
We blank the types it tries to parse so that it never attempts to parse.
This might be a problem if we want to create other pages, but for the API
end point it is fine.